requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.8.14
Django==5.2.1
django-apscheduler==0.7.0
django-jazzmin==3.0.1
django-rest-swagger==2.2.0
django-tenants==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==21.2.0
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
whitenoise==6.9.0
pytest==8.4.1
pytest-django==4.11.1
